requirements.txt
annotated-types==0.7.0
biopython==1.85
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
docopt-ng==0.9.0
et_xmlfile==2.0.0
ftpretty==0.4.0
goatools==1.5.2
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.3
obonet==1.1.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
patsy==1.0.1
platformdirs==4.5.0
pydantic==2.12.0
pydantic_core==2.41.1
pydot==4.0.1
Pygments==2.19.2
pyhpo==4.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytokens==0.1.10
pytz==2025.2
requests==2.32.5
rich==14.2.0
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
statsmodels==0.14.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.9
